tests/requirements.txt
pytest==3.5.1
PyYAML==3.12
requests==2.18.4
psycopg2==2.7.4